requirements.txt
assertpy==1.1
braceexpand==0.1.7
brotlipy @ file:///home/conda/feedstock_root/build_artifacts/brotlipy_1648854164373/work
certifi==2021.10.8
cffi @ file:///home/conda/feedstock_root/build_artifacts/cffi_1636046055389/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1644853463426/work
click==8.1.3
compressai==1.2.0
cryptography @ file:///home/conda/feedstock_root/build_artifacts/cryptography_1649035236822/work
cupy @ file:///home/conda/feedstock_root/build_artifacts/cupy_1651070280321/work
cycler==0.11.0
docker-pycreds==0.4.0
fastargs==1.2.0
fastrlock==0.8
ffcv==0.0.3
fonttools==4.33.3
gitdb==4.0.9
GitPython==3.1.27
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1642433548627/work
imageio==2.19.0
imageio-ffmpeg==0.4.7
imgcat==0.5.0
joblib==1.1.0
kiwisolver==1.4.2
llvmlite==0.38.0
matplotlib==3.5.1
networkx==2.8
numba @ file:///home/conda/feedstock_root/build_artifacts/numba_1644003844419/work
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1649806303256/work
opencv-python==4.5.5
packaging==21.3
pandas==1.4.2
pathtools==0.1.2
Pillow @ file:///home/conda/feedstock_root/build_artifacts/pillow_1649817996082/work
promise==2.3
protobuf==3.20.1
psutil==5.9.0
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1636257122734/work
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1643496850550/work
pyparsing==3.0.8
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1648857263093/work
python-dateutil==2.8.2
pytorch-msssim==0.2.1
pytorch-pfn-extras==0.5.8
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1641580202195/work
scikit-image==0.19.2
scikit-learn==1.0.2
scikit-video==1.1.11
scipy==1.8.0
sentry-sdk==1.5.10
setproctitle==1.2.3
shortuuid==1.0.8
six==1.16.0
sklearn==0.0
smmap==5.0.0
terminaltables==3.1.10
threadpoolctl==3.1.0
tifffile==2022.4.28
torch==1.11.0
torchvision==0.12.0
tqdm==4.64.0
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1650370875435/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1647489083693/work
wandb==0.12.15
webdataset==0.2.5
